feat(demo): allow client API keys
